+++ custom_components/keba_keenergy/number.py
@@ -9,6 +9,7 @@ from homeassistant.components.number import NumberDeviceClass
 from homeassistant.components.number import NumberEntity
 from homeassistant.components.number import NumberEntityDescription
 from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import PERCENTAGE
 from homeassistant.const import UnitOfTemperature
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
@@ -102,6 +103,17 @@ NUMBER_TYPES: dict[str, tuple[KebaKeEnergyNumberEntityDescription, ...]] = {
             },
             icon="mdi:thermometer-water",
             scale=1,
+        ),
+    ),
+    SectionPrefix.HEAT_PUMP: (
+        KebaKeEnergyNumberEntityDescription(
+            device_class=NumberDeviceClass.SPEED,
+            key="compressor_night_speed",
+            key_index=None,
+            native_unit_of_measurement=PERCENTAGE,
+            native_step=1,
+            translation_key="compressor_night_speed",
+            scale=100,
         ),
     ),
     SectionPrefix.BUFFER_TANK: (